requirements.txt
streamlit
requests
beautifulsoup4
gspread
oauth2client
trimesh
numpy
google-generativeai
